EasyCRM/requirements.txt
dj-database-url==0.5.0
Django==2.2
django-templated-mail==1.1.1
djangorestframework==3.12.1
djoser==2.0.5
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
psycopg2==2.8.6
